requirements.txt
dearpygui>=1.10.0
pyserial>=3.5
anthropic>=0.18.0  # Optional: for advanced AI diagnostic features
